linux sdk fixes: qt6keychain, libkml, linking order
qt6keychain on linux was using libsecret which had the issue that
it was using system libsecret instead of the one from vcpkg, but
we do not really need it anyway - and we can then also skip linking
to Glib which was causing frequent crashes upon app start for me.

app build was then failing when trying to link to libkml, which
is not included in our build, so now it is explicitly disabled
in gdal's portfile (should not be a problem because there's another
KML driver for gdal/ogr that does not rely on libkml).

the linking order for the sdk test was slightly incorrect for
geos and minizip, giving me link errors (spatialite needs geos,
freexl needs minizip)