Add libpq to overlays (to fix postgres crash upon project load)
In our current vcpkg baseline, there is libpq 16.0 and openssl 3.2.1,
and they do not work together well: connecting to postgres servers
fails (claiming that username was not provided, even if it was) and
qgis postgres provider was causing a crash because of that.

This adds libpq 16.2 port (which works fin with openssl 3.2.x)
from today's vcpkg (with no changes). Once we move to a newer vcpkg
baseline, this port can be safely removed.